// src/generated/client/models/Office.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Vehicle Management Services (management)
 * Vehicle Management Services (management)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Employee's office. Salary group should be set to VPLogistics if the office is KOTKA
 * @export
 */
export const Office = {
    Kotka: 'KOTKA',
    Kouvola: 'KOUVOLA',
    Rauha: 'RAUHA'
} as const;
export type Office = typeof Office[keyof typeof Office];


export function OfficeFromJSON(json: any): Office {
    return OfficeFromJSONTyped(json, false);
}

export function OfficeFromJSONTyped(json: any, ignoreDiscriminator: boolean): Office {
    return json as Office;
}

export function OfficeToJSON(value?: Office | null): any {
    return value as any;
}