// src/generated/client/models/PerDiemAllowanceType.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Vehicle Management Services (management)
 * Vehicle Management Services (management)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Per diem allowance
 * @export
 */
export const PerDiemAllowanceType = {
    Partial: 'PARTIAL',
    Full: 'FULL'
} as const;
export type PerDiemAllowanceType = typeof PerDiemAllowanceType[keyof typeof PerDiemAllowanceType];


export function PerDiemAllowanceTypeFromJSON(json: any): PerDiemAllowanceType {
    return PerDiemAllowanceTypeFromJSONTyped(json, false);
}

export function PerDiemAllowanceTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): PerDiemAllowanceType {
    return json as PerDiemAllowanceType;
}

export function PerDiemAllowanceTypeToJSON(value?: PerDiemAllowanceType | null): any {
    return value as any;
}