// src/generated/client/models/TruckDriveStateEnum.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Vehicle Management Services (management)
 * Vehicle Management Services (management)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TruckDriveStateEnum = {
    Rest: 'REST',
    DriverAvailable: 'DRIVER_AVAILABLE',
    Work: 'WORK',
    Drive: 'DRIVE',
    Error: 'ERROR',
    NotAvailable: 'NOT_AVAILABLE'
} as const;
export type TruckDriveStateEnum = typeof TruckDriveStateEnum[keyof typeof TruckDriveStateEnum];


export function TruckDriveStateEnumFromJSON(json: any): TruckDriveStateEnum {
    return TruckDriveStateEnumFromJSONTyped(json, false);
}

export function TruckDriveStateEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): TruckDriveStateEnum {
    return json as TruckDriveStateEnum;
}

export function TruckDriveStateEnumToJSON(value?: TruckDriveStateEnum | null): any {
    return value as any;
}